+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The file format of it is based on [Keep a Changelog](http://keepachangelog.com/e
 For public Changelog covering all changes done to Pipedrive’s API, webhooks and app extensions platforms, see [public Changelog](https://pipedrive.readme.io/docs/changelog) with discussion area in [Developers Community](https://devcommunity.pipedrive.com/c/documentation/changelog/19).
 
 ## [Unreleased]
+### Changed
+- Changed `POST /v1/users` request body type from `application/x-www-form-urlencoded` to `application/json`
+- Extracted `UserAccess` schema for defining the `access` property in `BaseUser` schema and in `POST /v1/users` endpoint
 
 ## 16.1.0
 ### Added